make policy an interface add option to always admit new items
