+++ ttl.go
@@ -122,7 +122,7 @@ func (m *expirationMap) del(key uint64, expiration time.Time) {
 // cleanup removes all the items in the bucket that was just completed. It deletes
 // those items from the store, and calls the onEvict function on those items.
 // This function is meant to be called periodically.
-func (m *expirationMap) cleanup(store *shardedMap, policy *lfuPolicy, onEvict itemCallback) {
+func (m *expirationMap) cleanup(store *shardedMap, policy policy, onEvict itemCallback) {
 	if m == nil {
 		return
 	}
